requirements.txt
python-dotenv==1.0.0
nada-dsl>=0.0.7
py-nillion-client>=0.0.7
pytest-asyncio>=0.23.6
seaborn==0.13.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4